+++ unnamed_engine/src/core/state.rs
@@ -8,10 +8,10 @@
 ////////////////////////////////////////////////////////////////////////////////
 // ? Where the data that will be used during execution is defined and handled.
 
-use wgpu::{util::DeviceExt, InstanceFlags};
+use wgpu::{util::DeviceExt, Device, InstanceFlags, RenderPipeline, Surface};
 use winit::window::Window;
 
-use crate::{renderer::{texture, camera::{Camera, CameraController}}, event::event::Event};
+use crate::{event::event::Event, renderer::{camera::{Camera, CameraController, CameraDescriptor, CameraUniform}, texture}};
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
@@ -53,25 +53,6 @@ const INDICES: &[u16] = &[
   2, 1, 3,
 ];
 
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct CameraUniform {
-  view_proj: [[f32; 4]; 4],
-}
-
-impl CameraUniform {
-  fn new() -> Self {
-    use cgmath::SquareMatrix;
-    Self {
-      view_proj: cgmath::Matrix4::identity().into(),
-    }
-  }
-
-  fn update_view_proj(&mut self, camera: &Camera) {
-    self.view_proj = camera.build_view_projection_matrix().into();
-  }
-}
-
 pub struct State {
   surface: wgpu::Surface,
   device: wgpu::Device,
@@ -84,10 +65,6 @@ pub struct State {
   index_buffer: wgpu::Buffer,
   num_indices: u32,
   diffuse_bind_group: wgpu::BindGroup,
-  camera: Camera,
-  camera_uniform: CameraUniform,
-  camera_buffer: wgpu::Buffer,
-  camera_bind_group: wgpu::BindGroup,
   camera_controller: CameraController,
   diffuse_texture: texture::Texture,
 }
@@ -192,52 +169,12 @@ impl State {
       }
     );
 
-    let camera = Camera {
-      eye: (0.0, 1.0, 2.0).into(),
-      target: (0.0, 0.0, 0.0).into(),
-      up: cgmath::Vector3::unit_y(),
+    let camera_controller = CameraController::new(&device, &CameraDescriptor {
+      speed: 0.2,
       aspect: config.width as f32 / config.height as f32,
       fovy: 45.0,
       near: 0.1,
-      far: 100.0,
-    };
-
-    let mut camera_uniform = CameraUniform::new();
-    camera_uniform.update_view_proj(&camera);
-
-    let camera_buffer = device.create_buffer_init(
-      &wgpu::util::BufferInitDescriptor {
-        label: Some("Camera Buffer"),
-        contents: bytemuck::cast_slice(&[camera_uniform]),
-        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-      }
-    );
-
-    let camera_bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-      entries: &[
-        wgpu::BindGroupLayoutEntry {
-          binding: 0,
-          visibility: wgpu::ShaderStages::VERTEX,
-          ty: wgpu::BindingType::Buffer {
-            ty: wgpu::BufferBindingType::Uniform,
-            has_dynamic_offset: false,
-            min_binding_size: None,
-          },
-          count: None,
-        },
-      ],
-      label: Some("camera_bind_group_layout"),
-    });
-
-    let camera_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-      layout: &camera_bind_group_layout,
-      entries: &[
-        wgpu::BindGroupEntry {
-          binding: 0,
-          resource: camera_buffer.as_entire_binding(),
-        }
-      ],
-      label: Some("camera_bind_group"),
+      far: 100.0
     });
 
     let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
@@ -249,7 +186,7 @@ impl State {
       label: Some("Render Pipeline Layout"),
       bind_group_layouts: &[
         &texture_bind_group_layout,
-        &camera_bind_group_layout,
+        &camera_controller.bind_group_layout,
       ],
       push_constant_ranges: &[],
     });
@@ -307,8 +244,6 @@ impl State {
 
     let num_indices = INDICES.len() as u32;
 
-    let camera_controller = CameraController::new(0.2);
-
     Self {
       window,
       surface,
@@ -322,10 +257,6 @@ impl State {
       num_indices,
       diffuse_texture,
       diffuse_bind_group,
-      camera,
-      camera_uniform,
-      camera_buffer,
-      camera_bind_group,
       camera_controller,
     }
   }
@@ -352,9 +283,8 @@ impl State {
   }
 
   pub fn update(&mut self) {
-    self.camera_controller.update_camera(&mut self.camera);
-    self.camera_uniform.update_view_proj(&self.camera);
-    self.queue.write_buffer(&self.camera_buffer, 0, bytemuck::cast_slice(&[self.camera_uniform]));
+    self.camera_controller.update();
+    self.queue.write_buffer(&self.camera_controller.buffer, 0, bytemuck::cast_slice(&[self.camera_controller.uniform]));
   }
 
   pub fn render(&mut self) -> Result<(), wgpu::SurfaceError> {
@@ -391,7 +321,7 @@ impl State {
 
       render_pass.set_pipeline(&self.render_pipeline);
       render_pass.set_bind_group(0, &self.diffuse_bind_group, &[]);
-      render_pass.set_bind_group(1, &self.camera_bind_group, &[]);
+      render_pass.set_bind_group(1, &self.camera_controller.bind_group, &[]);
       render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
       render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
       render_pass.draw_indexed(0..self.num_indices, 0, 0..1);