+++ unnamed_engine/src/renderer/camera.rs
@@ -1,3 +1,4 @@
+use wgpu::{util::DeviceExt, BindGroup, BindGroupLayout, Buffer, Device};
 ////////////////////////////////////////////////////////////////////////////////
 //                ██╗   ██╗███╗   ██╗███████╗███╗   ██╗                       //
 //                ██║   ██║████╗  ██║██╔════╝████╗  ██║                       //
@@ -8,7 +9,26 @@
 ////////////////////////////////////////////////////////////////////////////////
 // ? Defines the camera.
 use winit::keyboard::KeyCode;
-use crate::event::event::Event;
+use crate::{core::state::State, event::event::Event};
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct CameraUniform {
+  view_proj: [[f32; 4]; 4],
+}
+
+impl CameraUniform {
+  pub fn new() -> Self {
+    use cgmath::SquareMatrix;
+    Self {
+      view_proj: cgmath::Matrix4::identity().into(),
+    }
+  }
+
+  pub fn update_view_proj(&mut self, camera: &Camera) {
+    self.view_proj = camera.build_view_projection_matrix().into();
+  }
+}
 
 pub struct Camera {
   pub eye: cgmath::Point3<f32>,
@@ -28,7 +48,6 @@ pub const OPENGL_TO_WGPU_MATRIX: cgmath::Matrix4<f32> = cgmath::Matrix4::new(
   0.0, 0.0, 0.0, 1.0,
 );
 
-
 impl Camera {
   pub fn build_view_projection_matrix(&self) -> cgmath::Matrix4<f32> {
     let view = cgmath::Matrix4::look_at_rh(self.eye, self.target, self.up);
@@ -43,16 +62,83 @@ pub struct CameraController {
   is_backward_pressed: bool,
   is_left_pressed: bool,
   is_right_pressed: bool,
+  camera: Camera,
+  pub uniform: CameraUniform,
+  pub buffer: Buffer,
+  pub bind_group_layout: BindGroupLayout,
+  pub bind_group: BindGroup,
+}
+
+pub struct CameraDescriptor {
+  pub speed: f32,
+  pub aspect: f32,
+  pub fovy: f32,
+  pub near: f32,
+  pub far: f32,
 }
 
 impl CameraController {
-  pub fn new(speed: f32) -> Self {
+  /// Creates a new camera controller and initializes a camera with the passed
+  /// descriptor
+  pub fn new(device: &Device, desc: &CameraDescriptor) -> Self {
+    let mut camera = Camera {
+      eye: (0.0, 1.0, 2.0).into(),
+      target: (0.0, 0.0, 0.0).into(),
+      up: cgmath::Vector3::unit_y(),
+      aspect: desc.aspect,
+      fovy: 45.0,
+      near: 0.1,
+      far: 100.0,
+    };
+
+    let mut uniform = CameraUniform::new();
+    uniform.update_view_proj(&camera);
+
+    let buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+      label: Some("Camera Buffer"),
+      contents: bytemuck::cast_slice(&[uniform]),
+      usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
+      }
+    );
+
+    let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+      label: Some("camera_bind_group_layout"),
+      entries: &[
+        wgpu::BindGroupLayoutEntry {
+          binding: 0,
+          visibility: wgpu::ShaderStages::VERTEX,
+          ty: wgpu::BindingType::Buffer {
+            ty: wgpu::BufferBindingType::Uniform,
+            has_dynamic_offset: false,
+            min_binding_size: None
+          },
+          count: None,
+        },
+      ],
+    });
+
+    let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
+      label: Some("camera_bind_group"),
+      layout: &bind_group_layout,
+      entries: &[
+        wgpu::BindGroupEntry {
+          binding: 0,
+          resource: buffer.as_entire_binding(),
+        }
+      ]
+    });
+
     Self {
-      speed,
+      speed: desc.speed,
       is_forward_pressed: false,
       is_backward_pressed: false,
       is_left_pressed: false,
       is_right_pressed: false,
+      camera,
+      uniform,
+      buffer,
+      bind_group_layout,
+      bind_group,
     }
   }
 
@@ -85,31 +171,33 @@ impl CameraController {
         },
         _ => false
       }
-    }
+  }
 
-  pub fn update_camera(&self, camera: &mut Camera) {
+  pub fn update(&mut self) {
     use cgmath::InnerSpace;
-    let forward = camera.target - camera.eye;
+    let forward = self.camera.target - self.camera.eye;
     let forward_norm = forward.normalize();
     let forward_mag = forward.magnitude();
 
     if self.is_forward_pressed && forward_mag > self.speed {
-      camera.eye += forward_norm * self.speed;
+      self.camera.eye += forward_norm * self.speed;
     }
     if self.is_backward_pressed {
-      camera.eye -= forward_norm * self.speed;
+      self.camera.eye -= forward_norm * self.speed;
     }
 
-    let right = forward_norm.cross(camera.up);
+    let right = forward_norm.cross(self.camera.up);
 
-    let forward = camera.target - camera.eye;
+    let forward = self.camera.target - self.camera.eye;
     let forward_mag = forward.magnitude();
 
     if self.is_right_pressed {
-      camera.eye = camera.target - (forward + right * self.speed).normalize() * forward_mag;
+      self.camera.eye = self.camera.target - (forward + right * self.speed).normalize() * forward_mag;
     }
     if self.is_left_pressed {
-      camera.eye = camera.target - (forward - right * self.speed).normalize() * forward_mag;
+      self.camera.eye = self.camera.target - (forward - right * self.speed).normalize() * forward_mag;
     }
+
+    self.uniform.update_view_proj(&self.camera);
   }
 }